refactoring

+++ backend/routes/OrderRoutes.js
@@ -18,16 +18,38 @@ orderRouter.post("/order", UserAuth, async (req, res) => {
     }
   });
 
+
 //get all
+  
+  orderRouter.get("/allorder", async (req, res) => {
+    const createdAt = req.query.date == 'desc' ? -1 : 1
+      try {
+        const orders = await Order.find().sort({createdAt: createdAt});
+                                              
+        res.status(200).json(orders);
+      } catch (err) {
+        res.status(500).json({
+          message: err.message
+        });
+      }
+    });
+    
+
 
 orderRouter.get("/order", async (req, res) => {
   const createdAt = req.query.date == 'desc' ? -1 : 1
+  const search = req.query.search
     try {
-      const orders = await Order.find()
-      .sort({createdAt: createdAt});
+      const orders = await Order.find({$or:[
+        {refId: {'$regex': search}},
+        {email: {'$regex': search}},
+        {payment: {'$regex': search}}
+      ]}).sort({createdAt:createdAt});
       res.status(200).json(orders);
     } catch (err) {
-      res.status(500).json(err);
+      res.status(500).json({
+        message: err.message
+      });
     }
   });
 
@@ -43,46 +65,73 @@ orderRouter.get("/order/:userId", async(req, res)=>{
 })
 
 // delete order by id
-orderRouter.delete("/order/:refId", Verify, async(req,res)=>{
+orderRouter.delete("/order/:refId", UserAuth, async(req,res)=>{
+ try {
   const delete_id = req.params.refId
   if(delete_id == null){
       res.status(400).json({
-          status: 'error',
+          status: 400,
           message: 'order id should be provided'
       })
   }else{
+    if (req.user.isAdmin == true || req.user.superAdmin == true) {
       const del_prod= await Order.findOne({
         refId: req.params.refId
-  }).deleteOne()
+      }).deleteOne()
+    }else{
+      res.status(403).json({
+        msg:"unAuthorized user"
+      })
+    } 
   }
+ } catch (error) {
+  res.status(500).json({
+    msg: error.message
+  });
+ }
 })
 
 // delete all orders
 
-orderRouter.delete("/order/", Verify, async(req, res)=>{
-  const deleter = await Order.deleteMany()
-  if (deleter ){
-    res.send('success')
+orderRouter.delete("/order/", UserAuth, async(req, res)=>{
+  try {
+    if(req.user.isAdmin == true || req.user.superAdmin == true){
+      const deleter = await Order.deleteMany()
+      if (deleter ){
+        res.status(200).json({msg: "deleted successfully"})
+      }
+    }else{
+        res.status(403).json({
+          msg : "unauthorized user"
+        })
+    }
+  } catch (error) {
+    res.status(500).json({
+      msg: error.message
+    });
   }
+
 })
 
 // update status
 
-orderRouter.patch("/order/status", async(req,res)=>{
+orderRouter.patch("/order/status",async(req,res)=>{
   try {
-    await Order.find({status: 'pending'}).updateMany({
-      $set: {
-        status : 'successful'
-      }
-    }).then((result)=>{
-      console.log(result);
-      res.json({status: 'okay'})
-  }).catch((fail)=>{
-      console.log(fail);
-      res.json({status: 'fail'})
-  })
+      await Order.find({status: 'pending'}).updateMany({
+        $set: {
+          status : 'successful'
+        }
+      }).then((result)=>{
+        console.log(result);
+        res.json({status: 'okay'})
+    }).catch((fail)=>{
+        console.log(fail);
+        res.json({status: 'fail'})
+    })
   } catch (error) {
-    res.status(500).json(error);
+    res.status(500).json({
+      msg: error.message
+    });
     
   }
 })
